+++ scim2-sdk-common/src/main/java/com/unboundid/scim2/common/utils/MapperFactory.java
@@ -24,8 +24,11 @@ import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.MapperFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
+import com.fasterxml.jackson.databind.module.SimpleModule;
 
+import java.util.Calendar;
 import java.util.Collections;
+import java.util.Date;
 import java.util.Map;
 
 /**
@@ -139,11 +142,15 @@ public class MapperFactory
 
     // Don't serialize POJO nulls as JSON nulls.
     mapper.setSerializationInclusion(JsonInclude.Include.NON_NULL);
-    mapper.configure(SerializationFeature.WRITE_NULL_MAP_VALUES, false);
 
-    // Only use ISO8601 format for dates.
+    // Only use xsd:dateTime format for dates.
     mapper.configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, false);
-    mapper.setDateFormat(new ScimDateFormat());
+    SimpleModule dateTimeModule = new SimpleModule();
+    dateTimeModule.addSerializer(Calendar.class, new CalendarSerializer());
+    dateTimeModule.addDeserializer(Calendar.class, new CalendarDeserializer());
+    dateTimeModule.addSerializer(Date.class, new DateSerializer());
+    dateTimeModule.addDeserializer(Date.class, new DateDeserializer());
+    mapper.registerModule(dateTimeModule);
 
     // Do not care about case when de-serializing POJOs.
     mapper.configure(MapperFeature.ACCEPT_CASE_INSENSITIVE_PROPERTIES, true);