+++ scim2-sdk-common/src/test/java/com/unboundid/scim2/common/GenericScimResourceObjectTest.java
@@ -22,8 +22,8 @@ import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.google.common.collect.Lists;
 import com.unboundid.scim2.common.exceptions.ScimException;
 import com.unboundid.scim2.common.types.Meta;
+import com.unboundid.scim2.common.utils.DateTimeUtils;
 import com.unboundid.scim2.common.utils.JsonUtils;
-import com.unboundid.scim2.common.utils.ScimDateFormat;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
@@ -378,12 +378,12 @@ public class GenericScimResourceObjectTest
     Assert.assertEquals(gsr.getDateValue(path1), value1);
     Assert.assertEquals(GenericScimResource.getDateFromJsonNode(gsr.getValue(path1)),
         value1);
-    Assert.assertEquals(gsr.getStringValue(path1), new ScimDateFormat().format(value1));
+    Assert.assertEquals(gsr.getStringValue(path1), DateTimeUtils.format(value1));
 
     Assert.assertEquals(gsr.getDateValue(path2), value2);
     Assert.assertEquals(GenericScimResource.getDateFromJsonNode(gsr.getValue(path2)),
         value2);
-    Assert.assertEquals(gsr.getStringValue(path2), new ScimDateFormat().format(value2));
+    Assert.assertEquals(gsr.getStringValue(path2), DateTimeUtils.format(value2));
 
     List<Date> list1 = gsr.addDateValues(path3,
         Lists.<Date>newArrayList(arrayValue1, arrayValue2)).