+++ scim2-sdk-common/src/test/java/com/unboundid/scim2/common/JsonUtilsTestCase.java
@@ -23,11 +23,11 @@ import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
-import com.fasterxml.jackson.databind.util.ISO8601Utils;
 import com.google.common.collect.ImmutableMap;
 import com.unboundid.scim2.common.exceptions.ScimException;
 import com.unboundid.scim2.common.filters.Filter;
 import com.unboundid.scim2.common.types.Name;
+import com.unboundid.scim2.common.utils.DateTimeUtils;
 import com.unboundid.scim2.common.utils.JsonUtils;
 import com.unboundid.scim2.common.utils.MapperFactory;
 import org.testng.Assert;
@@ -35,7 +35,6 @@ import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
 import java.io.IOException;
-import java.text.ParsePosition;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -518,8 +517,8 @@ public class JsonUtilsTestCase
         Path.fromString("date"), gso.getObjectNode());
     assertEquals(dateResult.size(), 1);
     assertEquals(
-        ISO8601Utils.parse(dateResult.get(0).textValue(), new ParsePosition(0)),
-        ISO8601Utils.parse("2015-02-27T11:28:39Z", new ParsePosition(0)));
+        DateTimeUtils.parse(dateResult.get(0).textValue()),
+        DateTimeUtils.parse("2015-02-27T11:28:39Z"));
 
 
     List<JsonNode> binaryResult = JsonUtils.findMatchingPaths(
@@ -1201,5 +1200,21 @@ public class JsonUtilsTestCase
     Assert.assertEquals(name.getFamilyName(), "Smith");
     Assert.assertEquals(name.getGivenName(), "Bob");
     Assert.assertEquals(name.getMiddleName(), "X");
+  }
+
+  /**
+   * Test that the SCIM 2 SDK ObjectMapper ignores null map values.
+   */
+  @Test
+  public void testNullMapValue()
+  {
+    Map<String, String> map = new HashMap<>();
+    map.put("hasValue", "value1");
+    map.put("isNull", null);
+    ObjectNode objectNode = JsonUtils.valueToNode(map);
+
+    Assert.assertFalse(objectNode.path("hasValue").isMissingNode());
+    Assert.assertEquals(objectNode.path("hasValue").textValue(), "value1");
+    Assert.assertTrue(objectNode.path("isNull").isMissingNode());
   }
 }