+++ CHANGELOG.md
@@ -2,7 +2,17 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## v2.2.1 - unreleased
+## v2.3.0 - unreleased
+Updated SCIM 2 DateTime handling code to rely on xsd:dateTime code provided by JAXB rather than on Jackson APIs that are now deprecated. A DateTimeUtils class has been added to perform conversions between SCIM 2 DateTime strings and Java Date and Calendar objects. The ScimDateFormat class is now deprecated and should not be used.
+
+
+## v2.2.2 - 2019-06-24
+Updated the Jackson dependencies to 2.9.9, which addresses a potential security issue found in earlier versions of that library.
+
+
+## v2.2.1 - 2019-03-11
+Updated the jackson-databind dependency to 2.7.9.5, which addresses a number of security issues found in earlier versions of that library.
+
 Maven POM changes: Removed unused dependencies, declared previously implicit dependencies, declared 'test' scope as appropriate, and added a dependencyManagement section to the scim2-parent POM.
 
 Fixed an issue with `GenericScimResource.replaceValue(String, Date)` wrapping date values in double quotes.